requirements.txt
agno
openai
httpx
fastapi
python-multipart
uvicorn[standard]
# Supabase migration (Phase 1 + Phase 2)
supabase>=2.0.0
psycopg2-binary>=2.9.0
